# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /concepts:
    post:
      operationId: generateConcept
      tags: [concepts]
      summary: Generate a concept without a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConceptGenerateInput"
      responses:
        "201":
          description: Generated concept
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Concept"
  /concepts/{conceptId}:
    get:
      operationId: getConcept
      tags: [concepts]
      summary: Get a concept
      parameters:
        - $ref: "#/components/parameters/ConceptId"
      responses:
        "200":
          description: Concept
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Concept"
        "404":
          description: Concept not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /concepts/{conceptId}/favorite:
    patch:
      operationId: toggleConceptFavorite
      tags: [concepts]
      summary: Set a concept favorite state
      parameters:
        - $ref: "#/components/parameters/ConceptId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteInput"
      responses:
        "200":
          description: Updated concept
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Concept"
        "404":
          description: Concept not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  parameters:
...
    ConceptId:
      name: conceptId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Brief:
      type: object
      properties:
        furnitureLayout:
          type: string
        lighting:
          type: string
        walls:
          type: string
        flooring:
          type: string
        fixtures:
          type: string
        budget:
          type: string
        mustKeep:
          type: string
      required:
        - furnitureLayout
        - lighting
        - walls
        - flooring
        - fixtures
        - budget
        - mustKeep
...
    Concept:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: ["string", "null"]
        title:
          type: string
        roomType:
          type: string
        style:
          type: string
        mode:
          type: string
          enum: [restyle, renovate]
        sourceImageUrl:
          type: string
        resultImageUrl:
          type: string
        status:
          type: string
          enum: [completed]
        isFavorite:
          type: boolean
        brief:
          $ref: "#/components/schemas/Brief"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - title
        - roomType
        - style
        - mode
        - sourceImageUrl
        - resultImageUrl
        - status
        - isFavorite
        - brief
        - createdAt
    ConceptGenerateInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        roomType:
          type: string
          minLength: 1
        style:
          type: string
          minLength: 1
        mode:
          type: string
          enum: [restyle, renovate]
        sourceImageUrl:
          type: string
        brief:
          $ref: "#/components/schemas/Brief"
      required:
        - roomType
        - style
        - mode
        - sourceImageUrl
        - brief
    FavoriteInput:
      type: object
      properties:
        isFavorite:
          type: boolean
      required:
        - isFavorite
...